More Features
If someone says hello there in any capacity, the Bot replies with 'General Kenobi'. Occasionally, the bot will reply to a user and say that it has the highground, if the user replies with 'you underestimate my power', the bot then replies. If the user doesn't respond in time (10 messages between) then the process gets reset